// quasarnode/vue/src/store/generated/abag/quasarnode/intergamm/module/types/osmosis/gamm/pool-models/balancer/balancerPool.d.ts
export interface SmoothWeightChangeParams {
    /**
     * The start time for beginning the weight change.
     * If a parameter change / pool instantiation leaves this blank,
     * it should be generated by the state_machine as the current time.
     */
    startTime: Date | undefined;
    /** Duration for the weights to change over */
    duration: Duration | undefined;
    /**
     * The initial pool weights. These are copied from the pool's settings
     * at the time of weight change instantiation.
     * The amount PoolAsset.token.amount field is ignored if present,
     * future type refactorings should just have a type with the denom & weight
     * here.
     */
    initialPoolWeights: PoolAsset[];
    /**
     * The target pool weights. The pool weights will change linearly with respect
     * to time between start_time, and start_time + duration. The amount
     * PoolAsset.token.amount field is ignored if present, future type
     * refactorings should just have a type with the denom & weight here.
     */
    targetPoolWeights: PoolAsset[];
}
/**
 * BalancerPoolParams defined the parameters that will be managed by the pool
 * governance in the future. This params are not managed by the chain
 * governance. Instead they will be managed by the token holders of the pool.
 * The pool's token holders are specified in future_pool_governor.
 */
export interface BalancerPoolParams {
    swapFee: string;
    exitFee: string;
    smoothWeightChangeParams: SmoothWeightChangeParams | undefined;
}
// ...
export declare const SmoothWeightChangeParams: {
    encode(message: SmoothWeightChangeParams, writer?: Writer): Writer;
    decode(input: Reader | Uint8Array, length?: number): SmoothWeightChangeParams;
    fromJSON(object: any): SmoothWeightChangeParams;
    toJSON(message: SmoothWeightChangeParams): unknown;
    fromPartial(object: DeepPartial<SmoothWeightChangeParams>): SmoothWeightChangeParams;
};
export declare const BalancerPoolParams: {
    encode(message: BalancerPoolParams, writer?: Writer): Writer;
    decode(input: Reader | Uint8Array, length?: number): BalancerPoolParams;
    fromJSON(object: any): BalancerPoolParams;
    toJSON(message: BalancerPoolParams): unknown;
    fromPartial(object: DeepPartial<BalancerPoolParams>): BalancerPoolParams;
};
export declare const BalancerPool: {
    encode(message: BalancerPool, writer?: Writer): Writer;
    decode(input: Reader | Uint8Array, length?: number): BalancerPool;
    fromJSON(object: any): BalancerPool;
    toJSON(message: BalancerPool): unknown;
    fromPartial(object: DeepPartial<BalancerPool>): BalancerPool;
};
declare type Builtin = Date | Function | Uint8Array | string | number | undefined;
export declare type DeepPartial<T> = T extends Builtin ? T : T extends Array<infer U> ? Array<DeepPartial<U>> : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>> : T extends {} ? {
    [K in keyof T]?: DeepPartial<T[K]>;
} : Partial<T>;